Setup FitTV branch for webos4tv
:Release Notes:
Add FitTV option to build configuration

:Detailed Notes:
Add FitTV option to build configuration

:Testing Performed:
Local Test

:QA Notes:
N/A

:Issues Addressed:
[WRO-13307] Setup FitTV branch for webos4tv

Reviewed-on: http://gpro.lge.com/c/webosose/edge-ai-computer-vision/+/336684
Tested-by: Seungchul Han <seungchul.han@lge.com>

+++ framework/bodyPoseEstimation/fittv/FitTvPersonCropOperation.cpp
@@ -80,8 +80,8 @@ BBox FitTvPersonCropOperation::fixBbox(const std::shared_ptr<NodeInput>& input, 
     int modelInputWidth = 192;
     int modelInputHeight = 256;
 
-    float scalex = std::max( bbox.width, 10.0f );
-    float scaley = std::max( bbox.height, 10.0f );
+    float scalex = std::max( static_cast<float>(bbox.width), 10.0f );
+    float scaley = std::max( static_cast<float>(bbox.height), 10.0f );
     float arrayExpand = 1.0f;
     if ( bbox.width / bbox.height < 0.66 )
     {
